+++ src/main/java/com/map/gaja/client/presentation/dto/request/NewClientRequest.java
@@ -1,7 +1,7 @@
 package com.map.gaja.client.presentation.dto.request;
 
-import com.map.gaja.client.presentation.dto.subdto.AddressDto;
-import com.map.gaja.client.presentation.dto.subdto.LocationDto;
+import com.map.gaja.client.presentation.dto.request.subdto.AddressDto;
+import com.map.gaja.client.presentation.dto.request.subdto.LocationDto;
 import lombok.*;
 import org.springframework.web.multipart.MultipartFile;
